Removing realm_config and router_config files
