--- router/router_config.go
@@ -1,6 +1,0 @@
-package router
-
-type RouterConfig struct {
-	RealmConfigs  []*RealmConfig `json:"realms"`
-	RealmTemplate *RealmConfig   `json:"realm_template"`
-}
